Create iwa.zip
This zip file can be unzipped in your NSWASM/Newspeak folder and served  from the WebIDE subfolder. It is similar to the "out" folder created by the build script but heavily rearranged to meet some of the IWA requirements.
Once it is running direct your browser to chrome://internal-wep-apps and the newspeak app should be available to register on the dashboard. (Alot of this is from two month old memory).